trash deleted perm

+++ app/Http/Controllers/TrashController.php
@@ -2,12 +2,18 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Afor;
 use App\Models\Spot;
 use App\Models\Trash;
+use App\Models\Ifinal;
 use App\Models\Minimal;
+use App\Models\Progress;
 use Illuminate\Http\Request;
 use App\Models\Investigation;
+use App\Models\InvestigationLog;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Storage;
 
 class TrashController extends Controller
 {
@@ -15,14 +21,14 @@ class TrashController extends Controller
      * Display a listing of the resource.
      */
 
-     //Operation
-    public function trashOperationIndex()   
+    //Operation
+    public function trashOperationIndex()
     {
         return view('admin.trash.operation.index', [
             'active' => 'Operation',
             'user' => Auth::user(),
             // 'minimals' => Minimal::all(),
-            'investigations' => Investigation::latest()->get(),
+            'investigations' => Afor::where('deleted_at', '!=', null)->latest()->get(),
             'spots' => Spot::all(),
         ]);
     }
@@ -35,24 +41,74 @@ class TrashController extends Controller
         $investigations->delete();
         return redirect()->back()->with('success', 'Operation Report deleted successfully');
     }
-     //Investigation
-     public function trashinvestigationIndex()   
-     {
+    //Investigation
+    public function trashinvestigationIndex()
+    {
         return view('admin.trash.investigation.index', [
             'active' => 'Investigation',
             'user' => Auth::user(),
             // 'minimals' => Minimal::all(),
-            'investigations' => Investigation::latest()->get(),
+            'investigations' => Investigation::where('deleted_at', '!=', null)->latest()->get(),
             'spots' => Spot::all(),
         ]);
-     }
+    }
 
-     public function trashInvestigationDelete($id)
+    public function trashInvestigationDelete(Request $request)
     {
-        $investigations = Investigation::findorFail($id);
+        // dd($request->all());
+        $validated = $request->validate([
+            'password' => 'required',
+            'id' => 'required'
+        ]);
 
-        $investigations->delete();
-        return redirect()->back()->with('success', 'Investigation Report deleted successfully');
+        $profile = $request->user();
+
+        if (Hash::check($validated['password'], $profile->password)) {
+
+            $investigations = Investigation::findorFail($request->input('id'));
+            $log = new InvestigationLog();
+            $log->fill([
+                'investigation_id' => $investigations->id,
+                'user_id' => auth()->user()->id,
+                'details' => "Permanently Deleted an Investigation with a subject of " . $investigations->subject . ", Created on " . $investigations->date,
+                'action' => "Delete",
+            ]);
+            $log->save();
+            if ($investigations->minimal != null) {
+                # code...
+                $minimal = Minimal::findOrFail($investigations->minimal->id);
+
+                if ($minimal->photos !== null) {
+                    $photos = explode(", ", $minimal->photos);
+                    // dd($photos);
+                    foreach ($photos as $photoToDelete) {
+                        if (Storage::disk('public')->exists('minimal/' . $photoToDelete)) {
+                            // dd("photo is found: " . $photoToDelete);
+                            try {
+                                Storage::disk('public')->delete('minimal/' . $photoToDelete);
+                            } catch (\Throwable $th) {
+                                abort(404, "Page not found");
+                            }
+                        }
+                    }
+                }
+                $minimal->delete();
+            } elseif ($investigations->spot != null) {
+                $spot = Spot::findOrFail($investigations->spot->id);
+                $spot->delete();
+            } elseif ($investigations->progress != null) {
+                $progress = Progress::findOrFail($investigations->progress->id);
+                $progress->delete();
+            } elseif ($investigations->final != null) {
+                $final = Ifinal::findOrFail($investigations->final->id);
+                $final->delete();
+            }
+
+            $investigations->delete();
+            return redirect()->back()->with('success', 'Investigation Report deleted Permanently');
+        }else{
+            return redirect()->back()->with('status', 'Password Incorrect!');
+        }
     }
     /**
      * Show the form for creating a new resource.
@@ -89,7 +145,7 @@ class TrashController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request )
+    public function update(Request $request)
     {
         //
     }
